data_enrichment/requirements.txt
requests==2.31.0
google-generativeai==0.5.4
ollama==0.2.0
